renamed ensureToken fn ,review comments addressed
